Call get_taxonomy for taxonomy object instead of typecasting array
`get_taxonomy_labels` expects a "real" taxonomy object, not some fake coerced array of `$args`. Doing a typecasting of the array is causing some very edge-case odd behaviour, and segfualts PHP 7.2 in somem situations. I'm not sure why this was done originally done this way, but it's not quite correct, as `args !== taxonomy_object`.